more comments and a graceful wait

+++ core/server/master/src/main/java/alluxio/master/metastore/rocks/RocksInodeStore.java
@@ -29,6 +29,7 @@ import alluxio.metrics.MetricKey;
 import alluxio.metrics.MetricsSystem;
 import alluxio.proto.meta.InodeMeta;
 import alluxio.resource.CloseableIterator;
+import alluxio.util.SleepUtils;
 import alluxio.util.io.PathUtils;
 
 import com.google.common.annotations.VisibleForTesting;
@@ -278,6 +279,9 @@ public class RocksInodeStore implements InodeStore {
   }
 
   private long getProperty(String rocksPropertyName) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try {
       return db().getAggregatedLongProperty(rocksPropertyName);
     } catch (RocksDBException e) {
@@ -288,6 +292,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public void remove(Long inodeId) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try {
       byte[] id = Longs.toByteArray(inodeId);
       db().delete(mInodesColumn.get(), mDisableWAL, id);
@@ -298,6 +305,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public void writeInode(MutableInode<?> inode) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try {
       db().put(mInodesColumn.get(), mDisableWAL, Longs.toByteArray(inode.getId()),
           inode.toProto().toByteArray());
@@ -318,6 +328,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public void addChild(long parentId, String childName, Long childId) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try {
       db().put(mEdgesColumn.get(), mDisableWAL, RocksUtils.toByteArray(parentId, childName),
           Longs.toByteArray(childId));
@@ -328,6 +341,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public void removeChild(long parentId, String name) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try {
       db().delete(mEdgesColumn.get(), mDisableWAL, RocksUtils.toByteArray(parentId, name));
     } catch (RocksDBException e) {
@@ -337,6 +353,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public Optional<MutableInode<?>> getMutable(long id, ReadOption option) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     byte[] inode;
     try {
       inode = db().get(mInodesColumn.get(), Longs.toByteArray(id));
@@ -355,6 +374,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public CloseableIterator<Long> getChildIds(Long inodeId, ReadOption option) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     RocksIterator iter = db().newIterator(mEdgesColumn.get(), mReadPrefixSameAsStart);
     // first seek to the correct bucket
     iter.seek(Longs.toByteArray(inodeId));
@@ -379,11 +401,15 @@ public class RocksInodeStore implements InodeStore {
     RocksIter rocksIter = new RocksIter(iter, prefix);
     Stream<Long> idStream = StreamSupport.stream(Spliterators
         .spliteratorUnknownSize(rocksIter, Spliterator.ORDERED), false);
+    // TODO(jiacheng): check mClosed
     return CloseableIterator.create(idStream.iterator(), (any) -> iter.close());
   }
 
   @Override
   public Optional<Long> getChildId(Long inodeId, String name, ReadOption option) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     byte[] id;
     try {
       id = db().get(mEdgesColumn.get(), RocksUtils.toByteArray(inodeId, name));
@@ -396,6 +422,7 @@ public class RocksInodeStore implements InodeStore {
     return Optional.of(Longs.fromByteArray(id));
   }
 
+  // TODO(jiacheng): double check this
   static class RocksIter implements Iterator<Long> {
 
     final RocksIterator mIter;
@@ -444,6 +471,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public Optional<Inode> getChild(Long inodeId, String name, ReadOption option) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     return getChildId(inodeId, name).flatMap(id -> {
       Optional<Inode> child = get(id);
       if (!child.isPresent()) {
@@ -456,6 +486,9 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public boolean hasChildren(InodeDirectoryView inode, ReadOption option) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try (RocksIterator iter = db().newIterator(mEdgesColumn.get(), mReadPrefixSameAsStart)) {
       iter.seek(Longs.toByteArray(inode.getId()));
       return iter.isValid();
@@ -468,7 +501,7 @@ public class RocksInodeStore implements InodeStore {
     try (RocksIterator iter = db().newIterator(mEdgesColumn.get(),
         mIteratorOption)) {
       iter.seekToFirst();
-      while (iter.isValid()) {
+      while (!mClosed.get() && iter.isValid()) {
         long parentId = RocksUtils.readLong(iter.key(), 0);
         String childName = new String(iter.key(), Longs.BYTES, iter.key().length - Longs.BYTES);
         long childId = Longs.fromByteArray(iter.value());
@@ -485,7 +518,7 @@ public class RocksInodeStore implements InodeStore {
     try (RocksIterator iter = db().newIterator(mInodesColumn.get(),
         mIteratorOption)) {
       iter.seekToFirst();
-      while (iter.isValid()) {
+      while (!mClosed.get() && iter.isValid()) {
         inodes.add(getMutable(Longs.fromByteArray(iter.key()), ReadOption.defaults()).get());
         iter.next();
       }
@@ -518,14 +551,21 @@ public class RocksInodeStore implements InodeStore {
 
   @Override
   public void writeToCheckpoint(OutputStream output) throws IOException, InterruptedException {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     mRocksStore.writeToCheckpoint(output);
   }
 
   @Override
   public void restoreFromCheckpoint(CheckpointInputStream input) throws IOException {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     mRocksStore.restoreFromCheckpoint(input);
   }
 
+  // TODO(jiacheng): check mClosed?
   private class RocksWriteBatch implements WriteBatch {
     private final org.rocksdb.WriteBatch mBatch = new org.rocksdb.WriteBatch();
 
@@ -586,6 +626,8 @@ public class RocksInodeStore implements InodeStore {
   public void close() {
     mClosed.set(true);
     LOG.info("Closing RocksInodeStore and recycling all RocksDB JNI objects");
+    // Sleep to wait for all concurrent readers to either complete or abort
+    SleepUtils.sleepMs(Configuration.getMs(PropertyKey.ROCKS_GRACEFUL_SHUTDOWN_TIMEOUT));
     mRocksStore.close();
     mDisableWAL.close();
     mReadPrefixSameAsStart.close();
@@ -604,6 +646,9 @@ public class RocksInodeStore implements InodeStore {
    *         for debugging purposes
    */
   public String toStringEntries() {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     StringBuilder sb = new StringBuilder();
     try (ReadOptions readOptions = new ReadOptions().setTotalOrderSeek(true);
         RocksIterator inodeIter = db().newIterator(mInodesColumn.get(), readOptions)) {
